# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LlmConfig {
  id: number;
  serverUrl: string;
  port: number;
  gpuEnabled: boolean;
  defaultModel: string;
  createdAt: string;
  updatedAt: string;
}

export interface LlmConfigInput {
  serverUrl: string;
  port: number;
  gpuEnabled: boolean;
  defaultModel: string;
}

export interface LlmStatus {
  online: boolean;
  serverHealth: string;
  /** @nullable */
  version?: string | null;
  modelsCount?: number;
  runningModels?: string[];
  /** @nullable */
  error?: string | null;
}

export interface OllamaModel {
  name: string;
  size: number;
  digest: string;
  modifiedAt: string;
  /** @nullable */
  parameterSize?: string | null;
  /** @nullable */
  quantizationLevel?: string | null;
  /** @nullable */
  family?: string | null;
}

export interface RunningModel {
  name: string;
  size: number;
  sizeVram: number;
  expiresAt: string;
}

export interface PullModelInput {
  name: string;
}

export interface PullModelResponse {
  success: boolean;
  message: string;
}

export interface DeleteModelResponse {
  success: boolean;
  message: string;
}

export interface ChatRequestMessage {
  role: string;
  content: string;
}

export interface ChatRequest {
  model: string;
  messages: ChatRequestMessage[];
  temperature?: number;
  useRag?: boolean;
}

export interface ChatResponse {
  content: string;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  totalDuration?: number | null;
  /** @nullable */
  evalCount?: number | null;
  /** @nullable */
  ragContext?: string | null;
}

export interface Conversation {
  id: number;
  title: string;
  model: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateConversationInput {
  title: string;
  model: string;
}

export interface ChatMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  /** @nullable */
  rating?: number | null;
  createdAt: string;
}

export interface AddMessageInput {
  role: string;
  content: string;
}

export interface RateMessageInput {
  rating: number;
}

export interface ModelProfile {
  id: number;
  name: string;
  baseModel: string;
  systemPrompt: string;
  temperature: number;
  topP: number;
  topK: number;
  contextLength: number;
  repeatPenalty: number;
  deployed: string;
  createdAt: string;
  updatedAt: string;
}

export interface ModelProfileInput {
  name: string;
  baseModel: string;
  systemPrompt?: string;
  temperature?: number;
  topP?: number;
  topK?: number;
  contextLength?: number;
  repeatPenalty?: number;
}

export interface DeployProfileResponse {
  success: boolean;
  message: string;
}

export interface TrainingDataEntry {
  id: number;
  inputText: string;
  outputText: string;
  systemPrompt: string;
  category: string;
  quality: number;
  source: string;
  createdAt: string;
}

export interface TrainingDataInput {
  inputText: string;
  outputText: string;
  systemPrompt?: string;
  category?: string;
  quality?: number;
}

export interface ExportTrainingDataInput {
  format: string;
  minQuality?: number;
  /** @nullable */
  category?: string | null;
}

export interface CollectTrainingInput {
  conversationId: number;
  minRating?: number;
}

export interface CollectTrainingResponse {
  collected: number;
  message: string;
}

export type TrainingStatsByCategory = { [key: string]: number };

export type TrainingStatsByQuality = { [key: string]: number };

export interface TrainingStats {
  totalEntries: number;
  byCategory: TrainingStatsByCategory;
  byQuality: TrainingStatsByQuality;
  avgQuality: number;
}

export interface RagDocument {
  id: number;
  title: string;
  category: string;
  chunksCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateDocumentInput {
  title: string;
  content: string;
  category?: string;
}

export interface SearchDocumentsInput {
  query: string;
  maxResults?: number;
  /** @nullable */
  category?: string | null;
}

export interface SearchResult {
  chunkId: number;
  documentId: number;
  documentTitle: string;
  content: string;
  relevance: number;
}

export type RagStatsByCategory = { [key: string]: number };

export interface RagStats {
  totalDocuments: number;
  totalChunks: number;
  byCategory: RagStatsByCategory;
}