# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /llm/models/pull:
    post:
      operationId: pullModel
      tags: [llm-models]
      summary: Pull a model from Ollama registry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PullModelInput"
      responses:
        "200":
          description: Pull status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PullModelResponse"
  /llm/models/{name}:
    delete:
      operationId: deleteModel
      tags: [llm-models]
      summary: Delete a model from Ollama
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Model deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteModelResponse"
...
  /llm/chat:
    post:
      operationId: sendChatMessage
      tags: [llm-proxy]
      summary: Send a chat completion request to Ollama
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: Chat completion response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
...
  /chat/conversations:
    get:
      operationId: listConversations
      tags: [chat]
      summary: List all chat conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createConversation
      tags: [chat]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConversationInput"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
  /chat/conversations/{id}:
    delete:
      operationId: deleteConversation
      tags: [chat]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /chat/conversations/{id}/messages:
    get:
      operationId: getMessages
      tags: [chat]
      summary: Get messages for a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: addMessage
      tags: [chat]
      summary: Add a message to a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddMessageInput"
      responses:
        "201":
          description: Added message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
  /chat/messages/{messageId}/rate:
    patch:
      operationId: rateMessage
      tags: [chat]
      summary: Rate a chat message
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RateMessageInput"
      responses:
        "200":
          description: Rated message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
...
components:
  schemas:
...
    PullModelInput:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    PullModelResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
    DeleteModelResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
    ChatRequest:
      type: object
      properties:
        model:
          type: string
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatRequestMessage"
        temperature:
          type: number
        useRag:
          type: boolean
      required:
        - model
        - messages
    ChatRequestMessage:
      type: object
      properties:
        role:
          type: string
        content:
          type: string
      required:
        - role
        - content
    ChatResponse:
      type: object
      properties:
        content:
          type: string
        model:
          type: ["string", "null"]
        totalDuration:
          type: ["number", "null"]
        evalCount:
          type: ["integer", "null"]
        ragContext:
          type: ["string", "null"]
      required:
        - content
    Conversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        model:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - model
        - createdAt
        - updatedAt
    CreateConversationInput:
      type: object
      properties:
        title:
          type: string
        model:
          type: string
      required:
        - title
        - model
    ChatMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        rating:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    AddMessageInput:
      type: object
      properties:
        role:
          type: string
        content:
          type: string
      required:
        - role
        - content
    RateMessageInput:
      type: object
      properties:
        rating:
          type: integer
      required:
        - rating
...