# lib/api-zod/src/generated/types/chatRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChatRequestMessage } from "./chatRequestMessage";

export interface ChatRequest {
  model: string;
  messages: ChatRequestMessage[];
  temperature?: number;
  useRag?: boolean;
}