# lib/api-zod/src/generated/types/chatResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ChatResponse {
  content: string;
  /** @nullable */
  model?: string | null;
  /** @nullable */
  totalDuration?: number | null;
  /** @nullable */
  evalCount?: number | null;
  /** @nullable */
  ragContext?: string | null;
}